Create Shape_shifter_solver.cpp
This code recursively solves a puzzle. Given a map of swords and shields, the goal is to make every sword become a shield. When a shape is read into the map, every spot the shape is read in on flips so that every sword becomes a shield and every shield becomes a sword